backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
requests
python-multipart
pydantic
PyYAML
aiosqlite
google-auth
google-auth-oauthlib
langfuse
openai
prisma
qdrant_client
sentence-transformers
pyarrow==20.0.0
playwright
aiohttp
rapidfuzz
boto3